refactor: 베지에 곡선 관리 로직을 각 객체로 이동
