+++ app/src/main/java/woowacourse/paint/PaintView.kt
@@ -10,7 +10,6 @@ import woowacourse.paint.model.DrawMode
 import woowacourse.paint.model.pen.Pen
 import woowacourse.paint.model.shape.Eraser
 import woowacourse.paint.model.shape.Line
-import woowacourse.paint.model.shape.Line.Companion.updateLastPoint
 import woowacourse.paint.model.shape.Oval
 import woowacourse.paint.model.shape.Rectangle
 import woowacourse.paint.model.shape.Shapes
@@ -87,8 +86,7 @@ class PaintView(
             DrawMode.ERASER -> {
                 val paint = pen.getPaint()
                 val addEraserLine = Eraser(paint)
-                shapes.add(addEraserLine)
-                Eraser.updateLastPoint(pointX, pointY)
+                shapes.add(addEraserLine, pointX, pointY)
                 addEraserLine.path.moveTo(pointX, pointY)
                 setLayerType(LAYER_TYPE_HARDWARE, null)
                 invalidate()
@@ -100,10 +98,7 @@ class PaintView(
         when (drawMode) {
             DrawMode.LINE -> {
                 val line = shapes.last() as Line
-                val nextX = (Line.lastX + pointX) / 2
-                val nextY = (Line.lastY + pointY) / 2
-                line.path.quadTo(Line.lastX, Line.lastY, nextX, nextY)
-                Line.updateLastPoint(pointX, pointY)
+                line.quadTo(pointX, pointY)
                 invalidate()
             }
             DrawMode.RECT -> {
@@ -120,10 +115,7 @@ class PaintView(
             }
             DrawMode.ERASER -> {
                 val eraser = shapes.last() as Eraser
-                val nextX = (Eraser.lastX + pointX) / 2
-                val nextY = (Eraser.lastY + pointY) / 2
-                eraser.path.quadTo(Eraser.lastX, Eraser.lastY, nextX, nextY)
-                Eraser.updateLastPoint(pointX, pointY)
+                eraser.quadTo(pointX, pointY)
                 invalidate()
             }
         }
@@ -132,8 +124,7 @@ class PaintView(
     private fun addLine(pointX: Float, pointY: Float): Line {
         val paint = pen.getPaint()
         val addLine = Line(paint)
-        shapes.add(addLine)
-        updateLastPoint(pointX, pointY)
+        shapes.add(addLine, pointX, pointY)
         return addLine
     }
 }
